+++ app/src/main/kotlin/me/xizzhu/android/joshua/strongnumber/StrongNumberViewModel.kt
@@ -23,8 +23,10 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.filterNotNull
+import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 import me.xizzhu.android.joshua.core.BibleReadingManager
+import me.xizzhu.android.joshua.core.Settings
 import me.xizzhu.android.joshua.core.SettingsManager
 import me.xizzhu.android.joshua.core.StrongNumber
 import me.xizzhu.android.joshua.core.StrongNumberManager
@@ -38,6 +40,7 @@ import me.xizzhu.android.joshua.ui.createTitleStyleSpan
 import me.xizzhu.android.joshua.ui.recyclerview.BaseItem
 import me.xizzhu.android.joshua.ui.recyclerview.TextItem
 import me.xizzhu.android.joshua.ui.recyclerview.TitleItem
+import me.xizzhu.android.joshua.ui.recyclerview.VersePreviewItem
 import me.xizzhu.android.joshua.ui.setSpan
 import me.xizzhu.android.joshua.ui.toCharSequence
 import me.xizzhu.android.joshua.utils.firstNotEmpty
@@ -45,6 +48,8 @@ import me.xizzhu.android.logger.Log
 import javax.inject.Inject
 
 class StrongNumberViewData(val items: List<BaseItem>)
+
+class PreviewViewData(val settings: Settings, val title: String, val items: List<BaseItem>, val currentPosition: Int)
 
 @HiltViewModel
 class StrongNumberViewModel @Inject constructor(
@@ -118,4 +123,16 @@ class StrongNumberViewModel @Inject constructor(
     fun saveCurrentVerseIndex(verseToOpen: VerseIndex): Flow<ViewData<Unit>> = viewData {
         bibleReadingManager.saveCurrentVerseIndex(verseToOpen)
     }.onFailure { Log.e(tag, "Failed to save current verse", it) }
+
+    fun loadVersesForPreview(verseIndex: VerseIndex): Flow<ViewData<PreviewViewData>> = viewData {
+        val currentTranslation = bibleReadingManager.currentTranslation().firstNotEmpty()
+        val items = bibleReadingManager.readVerses(currentTranslation, verseIndex.bookIndex, verseIndex.chapterIndex)
+                .map { VersePreviewItem(it) }
+        PreviewViewData(
+                settings = settings().first(),
+                title = "${bibleReadingManager.readBookShortNames(currentTranslation)[verseIndex.bookIndex]}, ${verseIndex.chapterIndex + 1}",
+                items = items,
+                currentPosition = verseIndex.verseIndex
+        )
+    }.onFailure { Log.e(tag, "Failed to load verses for preview", it) }
 }